Add graphite_protcol parameter to config class
Under the Puppet future parser/4.0 language, variables used in
templates can only come from the local scope, and/or be explicitly
scoped with the scope.lookupvar function.

This adds a graphite_protocol parameter to the config class so that
the @graphite_protocol variable referenced in diamond.conf.erb will
be in the local scope.  Otherwise diamond.conf gets generated with
a blank GraphiteHandler/proto setting (when using the Puppet
future parser.)