hack: dont include python module
causes duplicate resources in fasrc repo given that we manage python
packages in a global package list and python puppet module tries to
install those same packages when it gets included. just using the
defined type without an include works because it avoids the install
phase in the python puppet module.